add init and delete

+++ include/fdf.h
@@ -6,7 +6,7 @@
 /*   By: aeclipso <aeclipso@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/29 11:00:30 by aeclipso          #+#    #+#             */
-/*   Updated: 2020/10/29 17:22:49 by aeclipso         ###   ########.fr       */
+/*   Updated: 2020/10/29 18:20:01 by aeclipso         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,8 +22,8 @@
 
 typedef struct	s_map
 {
-	int			height;
-	int			width;
+	size_t		height;
+	size_t		width;
 	int			**map;
 }				t_map;
